backend/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
google-genai
python-dotenv